requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.5.2
exceptiongroup==1.2.2
fastapi==0.115.4
fsspec==2024.9.0
h11==0.14.0
hpack==4.0.0
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.77
nvidia-nvtx-cu12==12.1.105
opencv-python==4.10.0.84
Pillow==9.3.0
protobuf==4.21.12
pydantic==2.9.2
pydantic_core==2.23.4
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.2
tensorboardX==2.6.2.2
torch==2.4.1
torchaudio==0.8.0a0+e4e171a
torchvision==0.19.1
triton==3.0.0
uvicorn==0.32.0
